Refactor: move everything to do with mall prices out of StoreManager into new MallPriceManager (#598)
