+++ src/net/sourceforge/kolmafia/textui/command/SearchMallCommand.java
@@ -3,7 +3,7 @@ package net.sourceforge.kolmafia.textui.command;
 import java.util.ArrayList;
 import java.util.List;
 import net.sourceforge.kolmafia.RequestLogger;
-import net.sourceforge.kolmafia.session.StoreManager;
+import net.sourceforge.kolmafia.session.MallPriceManager;
 import net.sourceforge.kolmafia.utilities.StringUtilities;
 
 public class SearchMallCommand extends AbstractCommand {
@@ -22,7 +22,7 @@ public class SearchMallCommand extends AbstractCommand {
       desiredLimit = StringUtilities.parseInt(splitup[1]);
     }
 
-    StoreManager.searchMall(parameters, desiredLimit, results);
+    MallPriceManager.searchMall(parameters, desiredLimit, results);
     RequestLogger.printList(results);
   }
 }
